simplify, remove useless code
