+++ .yarn/versions/a859e3b4.yml
@@ -1,0 +1,2 @@
+releases:
+  ts-overrides-plugin: patch
