+++ packages/plugin/src/cli/index.ts
@@ -1,13 +1,69 @@
 import * as path from 'node:path';
+import outmatch from 'outmatch';
 import type { PluginConfig, ProgramTransformer } from 'ts-patch';
 import type ts from 'typescript';
 
 import { type Override } from '../types/Override';
-import { getDiagnosticForFile, getDiagnosticsForProject, getOverridePrograms } from './utils';
 
 interface CliPluginConfig extends PluginConfig {
 	overrides: Override[];
 }
+
+export const getOverridePrograms = (
+	rootPath: string,
+	typescript: typeof ts,
+	overridesFromConfig: Override[],
+	rootFileNames: readonly string[],
+	defaultCompilerOptions: ts.CompilerOptions,
+	host?: ts.CompilerHost,
+): ts.Program[] => {
+	const dTsFiles = rootFileNames.filter(fileName => fileName.endsWith(`.d.ts`));
+
+	return overridesFromConfig.map(override => {
+		const isMatch = outmatch(override.files);
+		const filesToCurrentOverrideDiagnostic: string[] = rootFileNames.filter(fileName =>
+			isMatch(path.relative(rootPath, fileName)),
+		);
+
+		return typescript.createProgram(
+			[...filesToCurrentOverrideDiagnostic, ...dTsFiles],
+			{
+				...defaultCompilerOptions,
+				...typescript.convertCompilerOptionsFromJson(override.compilerOptions, rootPath).options,
+			},
+			host,
+		);
+	});
+};
+
+export const getDiagnosticForFile = (
+	overridePrograms: ts.Program[],
+	target: ts.Program,
+	sourceFile: ts.SourceFile,
+	method: 'getSemanticDiagnostics' | 'getBindAndCheckDiagnostics',
+	cancellationToken?: ts.CancellationToken,
+): readonly ts.Diagnostic[] => {
+	const { fileName } = sourceFile;
+
+	const overrideProgramForFile = overridePrograms.find(overrideProgram =>
+		overrideProgram.getRootFileNames().includes(fileName),
+	);
+
+	return overrideProgramForFile
+		? overrideProgramForFile[method](sourceFile, cancellationToken)
+		: target[method](sourceFile, cancellationToken);
+};
+
+export const getDiagnosticsForProject = (
+	program: ts.Program,
+	overridePrograms: ts.Program[],
+	cancellationToken?: ts.CancellationToken,
+): ts.Diagnostic[] =>
+	program
+		.getSourceFiles()
+		.flatMap(sourceFile =>
+			getDiagnosticForFile(overridePrograms, program, sourceFile, `getSemanticDiagnostics`, cancellationToken),
+		);
 
 const plugin: ProgramTransformer = (program, host, pluginConfig, extras) => {
 	const { overrides: overridesFromConfig } = pluginConfig as CliPluginConfig;
